behavior_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
tensorflow==2.14.0
pandas==2.1.2
django-cors-headers==4.3.0
